How this test is special?
